requirements.txt
streamlit>=1.28.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9

